refactor: split model.rs into focused modules (#23)
- Create model/ directory with proper module structure
- Move traits to model/traits.rs (DataPointBuilder, MetricCollector)
- Move enums to model/types.rs (Measurement, Unit, category enums)
- Move metric structs to model/metrics.rs with their implementations
- Move business logic to model/utilities.rs (batch_collect_metrics)
- Move test-only FailingDataPointBuilder to test_utils/mocks/collectors.rs
- Maintain backward compatibility through re-exports in model/mod.rs
- Fix visibility of PowerStatusBreakdownMetricCategory for proper re-export

This refactoring improves code organization by separating concerns into
focused modules, making the codebase more maintainable and easier to
navigate.
